refactor: make `seed/index.js` not self-executable
In order to be able to use `seeding` feature in the tests, now
`seed/index.js` exports some of its methods and `json` file with dummy
pre-defined users. The main function `seed` could be executed by `npm`
by the same `npm run seed` script. But now there's a `seed/fire.js` file
in the middle of this process.